캡스톤/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
yfinance==0.2.32
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.13.0
